requirements/base.txt
anyjson==0.3.3
billiard==3.3.0.20
celery==3.1.18
cffi==1.6.0
cryptography==1.1
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.4
django-allauth==0.19.1
django-annoying==0.8.1
django-appconf==1.0.1
django-bower==5.0.4
django-compressor==1.5
django-cors-headers==1.1.0
django-crispy-forms
django-custom-user==0.5
django-libsass==0.1
django-rest-auth==0.4.0
django-user-sessions==1.3.0
djangorestframework==3.1.1
django-session-idle-timeout==1.4.2
django-stripe-payments==2.0b34
django-toolbelt==0.0.1
django-extensions==1.2.0
ecdsa==0.13
Fabric==1.10.1
geoip2==2.3.0
httplib2==0.9.2
idna==2.1
importlib==1.0.3
ipaddr==2.1.11
ipaddress==1.0.14
kombu==3.0.26
libsass==0.8.2
oauthlib==0.7.2
paho-mqtt==1.1
paramiko==1.15.2
Pillow==2.8.1
pycrypto==2.6.1
python-openid==2.2.5
pytz==2016.4
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
validate-email==1.3
gunicorn==19.3.0
itsdangerous==0.24
mandrill==1.0.57
MarkupSafe==0.23
python-gcm==0.4
twilio==4.4.0
Werkzeug==0.10.4
wheel==0.24.0
WTForms==2.0.2
psycopg2==2.6.1
apns==2.0.1
django-widget-tweaks==1.4.1
